Simplify simulation interface and add notion of 'aging'
Replace the simulate_for overloads with the idea of the simulation being
'aged', where age_by(Minutes) is used to progress the simulation, and
age() is used to query how long (in virtual/simulated time) it has been
running. Minutes is now a conditionally-chosen type: Real if compiling
for CUDA, else the type equivalent to std::chrono::minutes, but with a
Real rather than integral representation. Identity semantics for
simulation objects are enforced by deleting copy-assignment and
copy-construction operators. The neighbor calculation algorithm has
been made more readable, to be eventually superseded by a configurable
algorithm for determining neighbors and thus implicitly defining the
shape of the simulation environment.